requirements_gui.txt
selenium>=4.15.0
Pillow>=10.0.0
webdriver-manager>=4.0.0
requests
pyinstaller>=5.0.0
